requirements.txt
torch==1.7.1
torchvision==0.8.2
tensorboard==2.6.0
pandas==1.1.5
opencv-python==4.5.1.48
scikit-learn==0.24.2
matplotlib==3.3.4
timm==0.4.12
tqdm==4.51.0
numpy==1.19.2
python-dotenv==0.19.0
Pillow==8.1.0
h5py==3.1.0
glob2==0.7